requirements.txt
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.0
certifi==2017.7.27.1
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.4.2
Django==1.11.6
django-allauth==0.33.0
django-anymail==1.0
django-cleanup==1.0.1
django-crispy-forms==1.6.1
django-environ==0.4.4
django-rest-framework==0.1.0
djangorestframework==3.7.0
gunicorn==19.7.1
idna==2.6
kombu==4.1.0
oauthlib==2.0.4
olefile==0.44
Pillow==4.3.0
psycopg2==2.7.3.1
python3-openid==3.1.0
pytz==2017.2
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
urllib3==1.22
vine==1.1.4
whitenoise==3.3.1
